refactor: Error Handling

+++ backend/src/controllers/card.controller.js
@@ -9,9 +9,9 @@ const create= async(req,res)=>{
 
     await cardServices.create(title,listId, boardId, user,(error,result)=>{
         if(error){
-            return res.status(500).send({message :error.message});
+            return res.status(500).send(error);
         }
-        return res.status(201).send(result);
+        return res.status(200).send(result);
     })
 }
 
@@ -19,24 +19,35 @@ const deleteById= async(req,res)=>{
     const user= req.user;
     const {boardId, listId, cardId}= req.params;
 
+    if(!(boardId && listId && cardId))
+        return res.status(400).send({message:"missing fields"});
+
     await cardServices.deleteById(cardId,listId,boardId,user,(error,result)=>{
         if(error)
-            return res.status(500).send({message:error.message})
+            return res.status(500).send(error)
 
         return res.status(200).send(result);
     })
 }
 
 const getCard= async(req,res)=>{
-    const user= req.user;
-    const {boardId, listId, cardId}= req.params;
+    try {
+        const user= req.user;
+        const {boardId, listId, cardId}= req.params;
 
-    await cardServices.getCard(cardId,listId,boardId,user,(error,result)=>{
-        if(error)
-            return res.status(500).send({message:error.message})
+        if (!boardId || !listId || !cardId) {
+            return res.status(400).send({ message: 'Missing required fields' });
+        }
 
-        return res.status(200).send(result);
+    
+        await cardServices.getCard(cardId,listId,boardId,user,(error,result)=>{
+            if(error)
+                return res.status(500).send(error)
+            return res.status(200).send(result);
     })
+    } catch (error) {
+        return res.status(500).send({ message: 'An unexpected error occurred', error: error.message });    }
+    
 }
 
 const updateCard= async(req,res)=>{
@@ -44,9 +55,12 @@ const updateCard= async(req,res)=>{
     const {boardId, listId, cardId}= req.params;
     const {title}= req.body; // add fields you want to update
 
+    if(!(boardId && listId && cardId))
+        return res.status(400).send({message:"missing fields"});
+
     await cardServices.update(cardId,listId,boardId,user,req.body,(error,result)=>{
         if(error)
-            return res.status(500).send({message:error.message})
+            return res.status(500).send(error)
 
         return res.status(200).send(result);
     })
@@ -57,6 +71,8 @@ const updateCover = async (req, res) => {
 	const { boardId, listId, cardId } = req.params;
 	const {color, isSizeOne} = req.body;
 	
+    if(!(boardId && listId && cardId))
+        return res.status(400).send({message:"missing fields"});
 
 	await cardServices.updateCover(
 		cardId,
